replace TicketingMessages with global message
