+++ src/travel-details-screens/DepartureDetailsScreenComponent.tsx
@@ -22,7 +22,6 @@ import {useGetServiceJourneyVehicles} from '@atb/travel-details-screens/use-get-
 import {StyleSheet, useTheme} from '@atb/theme';
 import {DepartureDetailsTexts, useTranslation} from '@atb/translations';
 import {TravelDetailsMapScreenParams} from '@atb/travel-details-map-screen/TravelDetailsMapScreenComponent';
-import {TicketingMessages} from '@atb/travel-details-screens/components/DetailsMessages';
 import {animateNextChange} from '@atb/utils/animation';
 import {formatToVerboseFullDate, isWithinSameDate} from '@atb/utils/date';
 import {getQuayName} from '@atb/utils/transportation-names';
@@ -44,6 +43,10 @@ import {Divider} from '@atb/components/divider';
 import {useMapData} from '@atb/travel-details-screens/use-map-data';
 import {useAnalytics} from '@atb/analytics';
 import {VehicleStatusEnumeration} from '@atb/api/types/generated/vehicles-types_v1';
+import {GlobalMessage, GlobalMessageContextEnum} from '@atb/global-messages';
+import {useRemoteConfig} from '@atb/RemoteConfigContext';
+import {useFirestoreConfiguration} from '@atb/configuration';
+import {canSellTicketsForSubMode} from '@atb/operator-config';
 
 export type DepartureDetailsScreenParams = {
   items: ServiceJourneyDeparture[];
@@ -64,6 +67,8 @@ export const DepartureDetailsScreenComponent = ({
   const [activeItemIndexState, setActiveItem] = useState(activeItemIndex);
   const {theme} = useTheme();
   const analytics = useAnalytics();
+  const {enable_ticketing} = useRemoteConfig();
+  const {modesWeSellTicketsFor} = useFirestoreConfiguration();
 
   const activeItem = items[activeItemIndexState];
   const hasMultipleItems = items.length > 1;
@@ -101,6 +106,17 @@ export const DepartureDetailsScreenComponent = ({
   const isJourneyFinished =
     vehiclePosition?.vehicleStatus === VehicleStatusEnumeration.Completed ||
     estimatedCallsWithMetadata.every((e) => e.actualArrivalTime);
+
+  const toQuay = estimatedCallsWithMetadata.find(
+    (estimatedCall) => estimatedCall.quay?.id === activeItem.toQuayId,
+  );
+  const fromQuay = estimatedCallsWithMetadata.find(
+    (estimatedCall) => estimatedCall.quay?.id === activeItem.fromQuayId,
+  );
+  const canSellTicketsForDeparture = canSellTicketsForSubMode(
+    subMode,
+    modesWeSellTicketsFor,
+  );
 
   const onPaginationPress = (newPage: number) => {
     animateNextChange();
@@ -247,11 +263,17 @@ export const DepartureDetailsScreenComponent = ({
             </View>
           )}
 
-          <TicketingMessages
-            item={items[0]}
-            trip={estimatedCallsWithMetadata}
-            mode={mode}
-            subMode={subMode}
+          <GlobalMessage
+            globalMessageContext={GlobalMessageContextEnum.appDepartureDetails}
+            ruleVariables={{
+              ticketingEnabled: enable_ticketing,
+              canSellTicketsForDeparture: canSellTicketsForDeparture,
+              mode: mode || null,
+              fromZones:
+                fromQuay?.quay?.tariffZones.map((zone) => zone.id) || null,
+              toZones: toQuay?.quay?.tariffZones.map((zone) => zone.id) || null,
+            }}
+            style={styles.messageBox}
           />
 
           <EstimatedCallRows
